More PS2KBW WIP
This approach isn't working. Whatever I had in my head for how the clock
work would in host->device mode, and using the 6522 shift registers with
that was all wrong. Looking at the scope it's clear we're 180 degrees
out of phase.

We're setting data on the rising edge of the clock, when the keyboard is
reading it. We need to be setting the data on the falling edge.

Going to abandon the SR approach and try bit-banging CB2 using CB1
interrupts. It's going to generate a little storm of interrupts, but the
host->device communications are pretty rare, so that's acceptable.